+++ src/node/index.ts
@@ -1,6 +1,6 @@
 import { UserFileConfig } from "./types.js";
+export const defineConfig = (config: UserFileConfig) => config;
 
 export * from "./start.js";
 export * from "./build.js";
-
-export const defineConfig = (config: UserFileConfig) => config;
+export { UserConfig } from "./types.js";