+++ src/client/components/Socials/styled.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const StyledSocialContainer = styled.div`
+  margin-left: auto;
+
+  svg + svg {
+    margin-left: 1rem;
+  }
+`;