+++ src/node/index.ts
@@ -1,2 +1,6 @@
-export * from './start.js';
-export * from './build.js';
+import { UserFileConfig } from "./types.js";
+export const defineConfig = (config: UserFileConfig) => config;
+
+export * from "./start.js";
+export * from "./build.js";
+export { UserConfig } from "./types.js";