+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add `ESM` info in `ReadMe` and `docs`
 - now support `socials` in `docit.config.js` file.
 - export more types
+- more docs
 
 
 ## [0.6.4] - 2022-03-29